+++ tests/testthat/test-allelematch_1-amDataset.R
@@ -1,12 +1,5 @@
 
-test_that("We are running the 3rd edition of testthat", code = {
-  # If this tests fails, then call
-  #   usethis::use_testthat(3)
-  # to configure DESCRIPTION to use 3rd edition of 'testthat'.
-  testthat::expect_gte(!!testthat::edition_get(), 3)
-})
-
-test_that("amExamples have not changed in allelematch", {
+test_that("allelematch::amExamples have not changed md5sum", {
 
     # Calculate a checksum for data stored under ./data/ in a package:
     env <- environment() # Use a local environment to avoid polluting the global environment with data sets
@@ -33,15 +26,16 @@ test_that("amExamples have not changed in allelematch", {
       # Only difference is the change in one column name from "gender" to "sex":
       expect_identical(md5sum("amExample5", package="allelematch"), '5f481f6287de5d5cc05277b645ba0642')
     }
+    gc() # Garbage collect the big data sets.
 
-    # Load the 2.5.5 version for use in our tests where we assume "gender":
-    data("amExample5", package="amregtest", envir = env) # Load the 2.5.5 version from amregtest
-
-    expect_identical(dim(amExample1), c( 20L, 22L))
-    expect_identical(dim(amExample2), c(148L, 22L))
-    expect_identical(dim(amExample3), c(319L, 22L))
-    expect_identical(dim(amExample4), c(307L, 22L))
-    expect_identical(dim(amExample5), c(335L, 23L))
+    # We use the 'amregtest' copies of the amExampleX data sets
+    # that match our tests. They are loaded once into
+    # the read-only environment `ro` in helper.R.
+    expect_identical(dim(ro$amExample1), c( 20L, 22L))
+    expect_identical(dim(ro$amExample2), c(148L, 22L))
+    expect_identical(dim(ro$amExample3), c(319L, 22L))
+    expect_identical(dim(ro$amExample4), c(307L, 22L))
+    expect_identical(dim(ro$amExample5), c(335L, 23L))
 })
 
 test_that("See how an object of class amDataset is built:", {
@@ -79,7 +73,7 @@ test_that("See how an object of class amDataset is built:", {
   }
 
   # Make amDataset with all optional parameters defaulted:
-  expect_snapshot(miniDataset1 <- amDataset(miniExample))
+  expect_snapshot(miniDataset1 <- amDataset(miniExample), variant = amvariant)
   expect_snapshot(print.amDataset(miniDataset1), variant = amvariant)
   {
     amDataset = miniDataset1
@@ -105,12 +99,13 @@ test_that("See how an object of class amDataset is built:", {
     expect_identical(!!amDataset$missingCode, "-99")
     expect_identical(!!amDataset$metadataColumn, NULL)
 
-    expect_snapshot_value(amDataset, style = "json2") # Stored under ./tests/testthat/_snaps/allelematch.md
+    expect_snapshot_value(amDataset, style = "json2", variant = amvariant)
   }
 
   # Make amDataset with all optional parameters set:
-  expect_snapshot(miniDataset2 <- amDataset(miniExample, missingCode="-88", indexColumn="sampleId", metaDataColumn="knownIndividual", ignoreColumn="dismiss."))
-  expect_snapshot(print.amDataset(miniDataset2), variant = amvariant)
+  expect_snapshot(miniDataset2 <- amDataset(miniExample, missingCode="-88", indexColumn="sampleId", metaDataColumn="knownIndividual", ignoreColumn="dismiss."),
+                  variant = amvariant)
+  expect_snapshot(print.amDataset(miniDataset2), variant = amvariant) # TODO 2.6.0
   {
     amDataset = miniDataset2
     expect_identical(!!class(amDataset), "amDataset")
@@ -128,18 +123,19 @@ test_that("See how an object of class amDataset is built:", {
     expect_identical(!!amDataset$metaData, c("A","A","B","  C")) # Spaces not stripped from metadata
     expect_identical(!!amDataset$missingCode, "-88")
 
-    expect_snapshot_value(amDataset, style = "json2")
+    expect_snapshot_value(amDataset, style = "json2", variant = amvariant)
   }
 
   # Make amDataset with all column parameters set as integers rather than characters:
-  expect_snapshot(miniDataset3 <- amDataset(miniExample, missingCode="-88", indexColumn=1, metaDataColumn=2, ignoreColumn=3))
+  expect_snapshot(miniDataset3 <- amDataset(miniExample, missingCode="-88", indexColumn=1, metaDataColumn=2, ignoreColumn=3),
+                  variant = amvariant)
   expect_snapshot(print.amDataset(miniDataset2), variant = amvariant)
   {
     # TODO : Catch none-character values for missingCode!!
     amDataset = miniDataset3
     expect_identical(amDataset, miniDataset2)
 
-    expect_snapshot_value(amDataset, style = "json2")
+    expect_snapshot_value(amDataset, style = "json2", variant = amvariant)
   }
 })
 
@@ -176,14 +172,15 @@ test_that("Different data types for arg to 'missingCode' give same result", {
     expect_identical(ds$missingCode, "NA")
     expect_type(ds$multilocus, "character")
     expect_identical(sum(unlist(ds$multilocus) == "NA"), 3L) # All 3 NA values now as strings
-    expect_snapshot_value(ds, style = "deparse", variant = amvariant) # style "json2" de-serializes "NA" to NA
+    # Using style "deparse" since style "json2" de-serializes "NA" to NA:
+    expect_snapshot_value(ds, style = "deparse", variant = amvariant) # TODO 2.6.0. Encoding of NA!
   }
 
   # Make sure arg missingCode = NA is converted to $missingCode="NA"
   expect_snapshot(ds2 <- amDataset(sample, missingCode = NA), variant = amvariant)
   {
     ds = ds2
-    expect_snapshot_value(ds, style = "deparse", variant = amvariant)  # style "json2" de-serializes "NA" to NA
+    expect_snapshot_value(ds, style = "deparse", variant = amvariant)
     expect_identical(ds$missingCode, "NA")
     expect_type(ds$multilocus, "character")
     expect_identical(sum(is.na(ds$multilocus)), 0L)