+++ tests/testthat/test-allelematch_6-amUnique.R
@@ -23,11 +23,11 @@ test_that("Loop the Loop", {
     "LOC2a"         = c(31:33, -99),
     "LOC2b"         = c(41:44)
   )
-  data("amExample1")
-  data("amExample2") ; amExample2 = amExample2[c(1:20),] # Just keep the first 20 rows to save speed and disk
-  data("amExample3") ; amExample3 = amExample3[c(1:20),] # Just keep the first 20 rows to save speed and disk
-  data("amExample4") ; amExample4 = amExample4[c(1:20),] # Just keep the first 20 rows to save speed and disk
-  data("amExample5") ; amExample5 = amExample5[c(1:20),] # Just keep the first 20 rows to save speed and disk
+  amExample1 <- ro$amExample1
+  amExample2 <- ro$amExample2[1:20, ] # Just keep the first 20 rows to save speed and disk
+  amExample3 <- ro$amExample3[1:20, ] # Just keep the first 20 rows to save speed and disk
+  amExample4 <- ro$amExample4[1:20, ] # Just keep the first 20 rows to save speed and disk
+  amExample5 <- ro$amExample5[1:20, ] # Just keep the first 20 rows to save speed and disk
 
   amdataMini     = amDataset(miniExample)
   amdataExample1 = amDataset(amExample1, indexColumn="sampleId", metaDataColumn="knownIndividual")
@@ -60,7 +60,7 @@ test_that("Loop the Loop", {
                       "\n  Rejected : ", cmdstr, "\n"))
 
         # Differ between expected and unexpected errors:
-        if (!grepl("no clusters formed.|'x' must be atomic", e_message, perl=TRUE)) {
+        if (!grepl("no clusters formed|'x' must be atomic", e_message, perl=TRUE)) {
           # Some unexpected error happened. Print it to the screen for easier debugging.
           message("\n  ", ret, sep="")
 
@@ -82,6 +82,7 @@ test_that("Loop the Loop", {
 
     return(ret)
   }
+
 
   # Parameters to amUnique:
   #  amUnique <- function(amDatasetFocal, multilocusMap=NULL, alleleMismatch=NULL, matchThreshold=NULL, cutHeight=NULL, doPsib="missing", consensusMethod=1, verbose=FALSE)