+++ tests/testthat/test-allelematch_6-amUnique_print.R
@@ -11,7 +11,7 @@ test_that("Print", {
     "LOC2a"         = c(31:33, -99),
     "LOC2b"         = c(41:44)
   )
-  data("amExample5") ; amExample5 = amExample5[c(1:20),] # Just keep the first 20 rows to save speed and disk
+  amExample5 <- ro$amExample5[1:20, ] # Just keep the first 20 rows to save speed and disk
 
   sink(nullfile()) # Drop cat output to stdout
   objMini     = amUnique(amDataset(miniExample), matchThreshold=0.5)
@@ -27,7 +27,7 @@ test_that("Print", {
   for (obj in c("objMini", "objExample5", "objExample5b", "objExample5c")) {
 
     # Write the name of the amUnique object to the _snap file:
-    expect_snapshot(paste("About to exercise", obj), variant=amvariant)
+    expect_snapshot_output(cat("About to exercise", obj), variant=amvariant)
 
     # allelematch:  Console summary is not available for "amUnique" objects.
     # Please use summary.amUnique(x, html=TRUE) or summary.amUnique(x, csv="file.csv") options.
@@ -43,18 +43,8 @@ test_that("Print", {
     tmp = tempfile(paste(obj, "_", sep=""), fileext=".html")
     expect_snapshot(summary.amUnique( get(obj), html=tmp),
                     variant=amvariant)
-    expect_snapshot(
-      cat(
-        sub("summary generated: </b><em>.+?</em>",
-            "summary generated: </b><em>(date)</em>",
-            gsub("(\\t| )+?(\\n|$)","\\2",
-                readLines(tmp, warn=FALSE),
-                perl=TRUE),
-            perl=TRUE),
-        sep="\n"),
-      variant=amvariant
-    )
-    file.remove(tmp)
+    snapshot_scrubHtmlFile(tmp, variant=amvariant)
+
   }