+++ tests/testthat/test-ggData.R
@@ -1,3 +1,4 @@
+
 # Run tests on [ggSample], a large real life data set.
 #
 # The tests are quite time consuming (approx one minute per test).